backend/delivery/requirements.txt
pika==1.3.2
python-dotenv==1.0.0
Flask==3.1.0
requests==2.28.1
